Collapse raw manifest code
